Clients Registered with Eureka(client cloud dependencies added)

+++ ratings-service/src/main/java/bejarano/ratingsservice/models/Rating.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class Rating {
 
-   private long recipeId;
+   private int recipeId;
    private int rating;
    private int time;
 }
